Removed libc dependency on Windows again, to fix building with Visual C++ 2005 Express Edition. Fixed performance problem with testsprite2 on the D3D driver.

+++ src/audio/SDL_audio.c
@@ -497,7 +497,7 @@ SDL_OpenAudio(SDL_AudioSpec * desired, SDL_AudioSpec * obtained)
     if (desired->channels == 0) {
         env = SDL_getenv("SDL_AUDIO_CHANNELS");
         if (env) {
-            desired->channels = SDL_atoi(env);
+            desired->channels = (Uint8)SDL_atoi(env);
         }
     }
     if (desired->channels == 0) {
@@ -517,7 +517,7 @@ SDL_OpenAudio(SDL_AudioSpec * desired, SDL_AudioSpec * obtained)
     if (desired->samples == 0) {
         env = SDL_getenv("SDL_AUDIO_SAMPLES");
         if (env) {
-            desired->samples = SDL_atoi(env);
+            desired->samples = (Uint16)SDL_atoi(env);
         }
     }
     if (desired->samples == 0) {
